requirements.txt
chatterbot==1.0.5
chatterbot_corpus==1.2.2
nltk==3.4.5
pymongo==3.10.1
pint==0.10
python-dateutil==2.7.5
pyyaml==5.1
spacy==2.1.8
tk